requirements.txt
imageio
numpy
opencv_python
Pillow
scikit_image
scipy
tqdm
